Rename pokemonData file for clarity
