+++ src/app/@modal/(.)pokedex/[id]/page.tsx
@@ -1,6 +1,6 @@
 import { Modal } from '@/components/Modal';
 import PokemonCard from '@/components/PokemonCard';
-import { fetchPokemonById } from '@/lib/pokemon';
+import { fetchPokemonById } from '@/lib/pokemonData';
 
 export default async function CardModal({
   params,
@@ -14,7 +14,7 @@ export default async function CardModal({
   if (!pokemon) return;
 
   return (
-    <Modal dialogClassName="[background-image:linear-gradient(-10deg,_#C97FE4,_#AECDF6)]">
+    <Modal dialogClassName="min-w-[40ch] [background-image:linear-gradient(-10deg,_#C97FE4,_#AECDF6)]">
       <PokemonCard pokemon={pokemon} />
     </Modal>
   );
